refactor(web): 抽象部分代码到独立文件
